pyproject.toml
[tool.poetry]
name = "new-e-commerce"
version = "0.1.0"
description = ""
authors = ["victro-nuzhniy <nuzhniyva@gmail.com>"]
readme = "README.md"
packages = [{include = "new_e_commerce"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
asyncpg = "0.25.0"
psycopg2-binary = "2.9.3"
python-dotenv = "^0.21.1"
pillow = "^9.4.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^22.10.0"
flake8-docstrings = "^1.6.0"
flake8 = "^5.0.4"
xenon = "^0.9.0"
isort = "^5.10.1"
mypy = "^0.990"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.20.2"
pytest-mock = "^3.10.0"
pytest-sugar = "^0.9.6"
pytest-cov = "^4.0.0"
pytest-randomly = "^3.12.0"
pytest-clarity = "^1.0.1"
faker = "^15.3.1"
pydantic-factories = "^1.7.0"
pytest-alembic = "^0.8.4"
httpx = "^0.23.0"
factory-boy = "^3.2.1"
pytest-freezegun = "^0.4.2"
pytest-django = "^4.5.2"
pytz = "^2022.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  | dist
  | .pytest_cache
  | .tox
  | docs
  | docs_src
  | .venv
  | .mypy_cache
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = []

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
# === Pytest ===
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = [".", " e_commerce"]
norecursedirs = ["venv", ".venv", ".env", "migrations"]
testpaths = ["tests",]
console_output_style = "progress"
python_files = "test_*"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

# === Pytest ===

